+++ custom_components/ecl_modbus/__init__.py
@@ -2,46 +2,104 @@ from __future__ import annotations
 
 """Home Assistant entry points for the ECL Modbus integration."""
 
+from typing import Any
+
 from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import CONF_NAME, CONF_PORT
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.typing import ConfigType
 
-from .const import DOMAIN, PLATFORMS
+from .const import (
+    DOMAIN,
+    PLATFORMS,
+    DEFAULT_BAUDRATE,
+    DEFAULT_NAME,
+    DEFAULT_SCAN_INTERVAL,
+    DEFAULT_SLAVE_ID,
+    CONF_BAUDRATE,
+    CONF_SCAN_INTERVAL,
+    CONF_SLAVE_ID,
+)
+from .modbus import EclModbusCoordinator, EclModbusHub
+from .registers import ALL_REGISTERS, option_key
+
+
+def _clamp_scan_interval(value: Any) -> int:
+    try:
+        sec = int(value)
+    except (TypeError, ValueError):
+        sec = DEFAULT_SCAN_INTERVAL
+    return max(5, min(3600, sec))
+
+
+def _default_enabled_key(reg_key: str) -> bool:
+    return reg_key in {"s3_temperature", "s4_temperature"}
+
+
+def _entry_store(hass: HomeAssistant, entry: ConfigEntry) -> dict[str, Any]:
+    domain_data = hass.data.setdefault(DOMAIN, {})
+    store = domain_data.get(entry.entry_id)
+    if not isinstance(store, dict):
+        store = {}
+        domain_data[entry.entry_id] = store
+    return store
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
-    """Basic setup from YAML (not used, but required by Home Assistant)."""
     return True
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    """Set up ECL Modbus from a config entry created via the UI."""
-    hass.data.setdefault(DOMAIN, {})
+    """Create hub+coordinator and forward to platforms."""
+    store = _entry_store(hass, entry)
 
-    # Forward the config entry to all platforms (sensor, number, etc.)
+    # Build enabled register list from options
+    enabled_regs = []
+    for reg in ALL_REGISTERS:
+        enabled = bool(
+            entry.options.get(option_key(reg.key), _default_enabled_key(reg.key))
+        )
+        if enabled:
+            enabled_regs.append(reg)
+
+    # Reuse hub if it already exists (prevents serial port lock issues on reload)
+    hub = store.get("hub")
+    if not isinstance(hub, EclModbusHub):
+        hub = EclModbusHub(
+            port=entry.data[CONF_PORT],
+            baudrate=entry.data.get(CONF_BAUDRATE, DEFAULT_BAUDRATE),
+            slave_id=entry.data.get(CONF_SLAVE_ID, DEFAULT_SLAVE_ID),
+        )
+        store["hub"] = hub
+
+    # Create a fresh coordinator every setup (uses current enabled_regs + scan_interval)
+    scan_interval_sec = _clamp_scan_interval(entry.options.get(CONF_SCAN_INTERVAL))
+    coordinator = EclModbusCoordinator(hass, hub, enabled_regs, scan_interval_sec)
+
+    # First refresh BEFORE platforms set up (prevents coordinator.data=None issues)
+    await coordinator.async_config_entry_first_refresh()
+
+    store["coordinator"] = coordinator
+    store["registers"] = enabled_regs
+    store["entry_title"] = entry.data.get(CONF_NAME, DEFAULT_NAME)
+
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
-
-    # Reload integration automatically when the entry is updated (options changed)
     entry.async_on_unload(entry.add_update_listener(async_reload_entry))
-
     return True
 
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    """Unload a config entry and close the Modbus client."""
-    # Close the Modbus hub (if any) to free the serial port
-    entry_store = hass.data.get(DOMAIN, {}).get(entry.entry_id)
+    store = hass.data.get(DOMAIN, {}).get(entry.entry_id, {})
+    hub = store.get("hub")
 
-    if isinstance(entry_store, dict):
-        hub = entry_store.get("hub")
-        if hub is not None:
-            try:
-                hub.close()
-            except Exception:  # noqa: BLE001
-                pass
-
-    # Unload all platforms
     unload_ok = await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
+
+    # Close hub AFTER platforms are unloaded (safer)
+    if hub is not None:
+        try:
+            hub.close()
+        except Exception:  # noqa: BLE001
+            pass
 
     if unload_ok and DOMAIN in hass.data:
         hass.data[DOMAIN].pop(entry.entry_id, None)
@@ -50,5 +108,4 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
 
 async def async_reload_entry(hass: HomeAssistant, entry: ConfigEntry) -> None:
-    """Handle config entry updates (e.g. options changed) by reloading."""
     await hass.config_entries.async_reload(entry.entry_id)
