+++ custom_components/ecl_modbus/const.py
@@ -3,27 +3,16 @@ from __future__ import annotations
 """
 Constants for the ECL Modbus integration.
 
-This file must contain ONLY static constants and configuration keys.
-No runtime logic, no register definitions, no enable flags.
+Keep ONLY constants and config keys here.
+No runtime logic.
 """
-
-# -----------------------------------------------------------------------------
-# Integration basics
-# -----------------------------------------------------------------------------
 
 DOMAIN = "ecl_modbus"
 
-# Platforms provided by this integration
-PLATFORMS: list[str] = [
-    "sensor",   # Read-only registers
-    "number",   # Writable numeric registers (RW)
-    # "select", # Future: token-based RW registers (Comfort/Night/etc.)
-]
+# Platforms we expose
+PLATFORMS: list[str] = ["sensor", "number"]
 
-# -----------------------------------------------------------------------------
-# Config flow base options
-# -----------------------------------------------------------------------------
-
+# Config entry (setup)
 CONF_BAUDRATE = "baudrate"
 CONF_SLAVE_ID = "slave_id"
 CONF_SCAN_INTERVAL = "scan_interval"
